refactor(worker): 将 ResourceStrategy 迁移为 HookAction，统一资源处理机制
- 将 strategy/resource/ 下 8 个策略子类重写为 hook/resource/ 下的 HookAction 实现
  (Download/Shell/ExecShell/Link/Replace/AppendLine/Frontend/NexusResource)
- 删除 ResourceStrategy 抽象类、死桥接器 ResourceStrategyHookAction 及其重复 SPI 缓存
- 删除 InstallServiceHandler 中的 static cache 与 execResourceStrategies() 方法
- YarnInstallServiceHandler.install() 简化为校验 hadoop 目录后直接返回成功
- WorkerCommandGrpcService 注入 frameCode 到 hook 变量，供 DownloadStrategy 使用
- 删除 InstallServiceRoleCommand / ServiceRoleInfo 中的 resourceStrategies 字段
- 删除 ServiceUpgradeHandler 中的 setResourceStrategies() 调用
- 将 21 个 service_ddl.json 的 resourceStrategies 条目转换为 POST_INSTALL hooks（共 69 条）
- 删除无引用的 EmptyStrategy
