+++ datasophon-api/src/main/resources/meta/datacluster/GRAFANA/service_ddl.json
@@ -7,7 +7,7 @@
   "dependencies": [],
   "packageName": "grafana-9.5.21.linux-amd64.tar.gz",
   "decompressPackageName": "grafana-v9.5.21",
-  "arch":{
+  "arch": {
     "x86_64": {
       "packageName": "grafana-9.5.21.linux-amd64.tar.gz"
     },
@@ -27,17 +27,6 @@
       "cardinality": 1,
       "logFile": "data/log/grafana.log",
       "jmxPort": "3000",
-      "resourceStrategies":[{
-        "type": "download",
-        "from": "script/control.sh",
-        "to": "control.sh",
-        "md5": "31179a30dd4d8c3314a25eb32ddc0c7f"
-      },{
-        "type": "download",
-        "from": "data/grafana.db",
-        "to": "data/grafana.db",
-        "md5": "4841ba20b7940a44f85899d0083287cb"
-      }],
       "startRunner": {
         "timeout": "60",
         "program": "control.sh",
@@ -74,33 +63,55 @@
         "name": "Grafana Ui",
         "label": "Grafana Ui",
         "url": "http://${host}:3000"
-      }
+      },
+      "hooks": [
+        {
+          "type": "POST_INSTALL",
+          "action": "download",
+          "params": {
+            "from": "script/control.sh",
+            "to": "control.sh",
+            "md5": "31179a30dd4d8c3314a25eb32ddc0c7f"
+          }
+        },
+        {
+          "type": "POST_INSTALL",
+          "action": "download",
+          "params": {
+            "from": "data/grafana.db",
+            "to": "data/grafana.db",
+            "md5": "4841ba20b7940a44f85899d0083287cb"
+          }
+        }
+      ]
     }
   ],
   "configWriter": {
-    "generators": [{
-       "filename": "custom.ini",
-       "configFormat": "custom",
-       "outputDirectory": "conf",
-       "templateName": "grafana.ftl",
-       "conditionalOnProperty": "datasophon.proxy-grafana.enable",
-       "includeParams": [
-          "clusterId"
-       ]
-    }]
-  },
-  "parameters": [
+    "generators": [
       {
-        "name": "clusterId",
-        "label": "集群ID",
-        "description": "集群ID",
-        "configType": "map",
-        "required": true,
-        "type": "input",
-        "value": "${clusterId}",
-        "configurableInWizard": false,
-        "hidden": true,
-        "defaultValue": "${clusterId}"
+        "filename": "custom.ini",
+        "configFormat": "custom",
+        "outputDirectory": "conf",
+        "templateName": "grafana.ftl",
+        "conditionalOnProperty": "datasophon.proxy-grafana.enable",
+        "includeParams": [
+          "clusterId"
+        ]
       }
     ]
+  },
+  "parameters": [
+    {
+      "name": "clusterId",
+      "label": "集群ID",
+      "description": "集群ID",
+      "configType": "map",
+      "required": true,
+      "type": "input",
+      "value": "${clusterId}",
+      "configurableInWizard": false,
+      "hidden": true,
+      "defaultValue": "${clusterId}"
+    }
+  ]
 }
