+++ datasophon-api/src/main/resources/meta/datacluster/JUICEFS/service_ddl.json
@@ -25,20 +25,6 @@
       "roleType": "worker",
       "cardinality": "1+",
       "logFile": "logs/appweb",
-      "resourceStrategies": [
-        {
-          "type": "download",
-          "from": "script/juicefs-format-wrapper.sh",
-          "to": "bin/juicefs-format-wrapper.sh",
-          "md5": "42afbf46e9625051f4637074fb7298ce"
-        },
-        {
-          "type": "execShell",
-          "commands": [
-            "${ROOT.JUICEFS.INSTALL_PATH}/bin/juicefs-format-wrapper.sh  --storage ${JUICEFS.juicefsStorage} --bucket ${JUICEFS.juicefsBucket}  --access-key ${JUICEFS.juicefsAccessKey}  --secret-key ${JUICEFS.juicefsSecretKey} --no-update '${JUICEFS.juicefsMeta}'  ${JUICEFS.juicefsFileSystemName}"
-          ]
-        }
-      ],
       "jmxPort": 9567,
       "startRunner": {
         "timeout": "60",
@@ -67,7 +53,27 @@
         "args": [
           "restart"
         ]
-      }
+      },
+      "hooks": [
+        {
+          "type": "POST_INSTALL",
+          "action": "download",
+          "params": {
+            "from": "script/juicefs-format-wrapper.sh",
+            "to": "bin/juicefs-format-wrapper.sh",
+            "md5": "42afbf46e9625051f4637074fb7298ce"
+          }
+        },
+        {
+          "type": "POST_INSTALL",
+          "action": "execShell",
+          "params": {
+            "commands": [
+              "${ROOT.JUICEFS.INSTALL_PATH}/bin/juicefs-format-wrapper.sh  --storage ${JUICEFS.juicefsStorage} --bucket ${JUICEFS.juicefsBucket}  --access-key ${JUICEFS.juicefsAccessKey}  --secret-key ${JUICEFS.juicefsSecretKey} --no-update '${JUICEFS.juicefsMeta}'  ${JUICEFS.juicefsFileSystemName}"
+            ]
+          }
+        }
+      ]
     }
   ],
   "configWriter": {
