+++ datasophon-api/src/main/resources/meta/datacluster/SPARK3/service_ddl.json
@@ -20,30 +20,39 @@
       },
       "cardinality": "1+",
       "logFile": "logs/hadoop-${user}-datanode-${host}.log",
-      "resourceStrategies": [
+      "hooks": [
         {
-          "type": "sh",
-          "commands": [
-            [
-              "mkdir",
-              "logs"
-            ],
-            [
-              "cp",
-              "/data/install_datasophon/datasophon-worker/lib/mysql-connector-java-8.0.28.jar",
-              "jars/"
+          "type": "POST_INSTALL",
+          "action": "sh",
+          "params": {
+            "commands": [
+              [
+                "mkdir",
+                "logs"
+              ],
+              [
+                "cp",
+                "/data/install_datasophon/datasophon-worker/lib/mysql-connector-java-8.0.28.jar",
+                "jars/"
+              ]
             ]
-          ]
+          }
         },
         {
-          "type": "nexus",
-          "from": "iceberg-spark-runtime-3.4_2.12-1.3.1.jar",
-          "to": "jars/iceberg-spark-runtime-3.4_2.12-1.3.1.jar"
+          "type": "POST_INSTALL",
+          "action": "nexus",
+          "params": {
+            "from": "iceberg-spark-runtime-3.4_2.12-1.3.1.jar",
+            "to": "jars/iceberg-spark-runtime-3.4_2.12-1.3.1.jar"
+          }
         },
         {
-          "type": "nexus",
-          "from": "spark-doris-connector-spark-3.4-25.2.0.jar",
-          "to": "jars/spark-doris-connector-spark-3.4-25.2.0.jar"
+          "type": "POST_INSTALL",
+          "action": "nexus",
+          "params": {
+            "from": "spark-doris-connector-spark-3.4-25.2.0.jar",
+            "to": "jars/spark-doris-connector-spark-3.4-25.2.0.jar"
+          }
         }
       ]
     }
@@ -136,23 +145,57 @@
       "required": false,
       "type": "multipleWithKey",
       "value": [
-        {"spark.sql.hive.metastore.jars": "/data/install_datasophon/hive/lib/*"},
-        {"spark.sql.hive.metastore.version": "3.1.3"},
-        {"spark.yarn.jars": "hdfs:///spark3/jars/*.jar"},
-        {"spark.port.maxRetries": "100"},
-        {"spark.sql.catalogImplementation": "hive"},
-        {"spark.sql.uri": "${HIVE.hive.metastore.uris}"},
-        {"spark.sql.extensions": "org.apache.iceberg.spark.extensions.IcebergSparkSessionExtensions,org.apache.paimon.spark.extensions.PaimonSparkSessionExtensions"},
-        {"spark.sql.warehouse.dir": "/user/hive/warehouse"},
-        {"spark.sql.catalog.paimon_catalog": "org.apache.paimon.spark.SparkCatalog"},
-        {"spark.sql.catalog.paimon_catalog.metastore": "hive"},
-        {"spark.sql.catalog.paimon_catalog.uri": "${HIVE.hive.metastore.uris}"},
-        {"spark.sql.catalog.paimon_catalog.warehouse": "hdfs:///user/hive/warehouse"},
-        {"spark.sql.catalog.doris_catalog": "org.apache.doris.spark.catalog.DorisTableCatalog"},
-        {"spark.sql.catalog.doris_catalog.doris.fenodes": "${DORIS.feMaster}:${DORIS.http_port}"},
-        {"spark.sql.catalog.doris_catalog.doris.query.port": "${DORIS.query_port}"},
-        {"spark.sql.catalog.doris_catalog.doris.user": "root"},
-        {"spark.sql.catalog.doris_catalog.doris.password": "${DORIS.root_password}"}
+        {
+          "spark.sql.hive.metastore.jars": "/data/install_datasophon/hive/lib/*"
+        },
+        {
+          "spark.sql.hive.metastore.version": "3.1.3"
+        },
+        {
+          "spark.yarn.jars": "hdfs:///spark3/jars/*.jar"
+        },
+        {
+          "spark.port.maxRetries": "100"
+        },
+        {
+          "spark.sql.catalogImplementation": "hive"
+        },
+        {
+          "spark.sql.uri": "${HIVE.hive.metastore.uris}"
+        },
+        {
+          "spark.sql.extensions": "org.apache.iceberg.spark.extensions.IcebergSparkSessionExtensions,org.apache.paimon.spark.extensions.PaimonSparkSessionExtensions"
+        },
+        {
+          "spark.sql.warehouse.dir": "/user/hive/warehouse"
+        },
+        {
+          "spark.sql.catalog.paimon_catalog": "org.apache.paimon.spark.SparkCatalog"
+        },
+        {
+          "spark.sql.catalog.paimon_catalog.metastore": "hive"
+        },
+        {
+          "spark.sql.catalog.paimon_catalog.uri": "${HIVE.hive.metastore.uris}"
+        },
+        {
+          "spark.sql.catalog.paimon_catalog.warehouse": "hdfs:///user/hive/warehouse"
+        },
+        {
+          "spark.sql.catalog.doris_catalog": "org.apache.doris.spark.catalog.DorisTableCatalog"
+        },
+        {
+          "spark.sql.catalog.doris_catalog.doris.fenodes": "${DORIS.feMaster}:${DORIS.http_port}"
+        },
+        {
+          "spark.sql.catalog.doris_catalog.doris.query.port": "${DORIS.query_port}"
+        },
+        {
+          "spark.sql.catalog.doris_catalog.doris.user": "root"
+        },
+        {
+          "spark.sql.catalog.doris_catalog.doris.password": "${DORIS.root_password}"
+        }
       ],
       "configurableInWizard": true,
       "hidden": false,
