+++ datasophon-worker/src/main/java/com/datasophon/worker/hook/resource/ReplaceStrategy.java
@@ -1,0 +1,53 @@
+package com.datasophon.worker.hook.resource;
+
+import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.bean.copier.CopyOptions;
+import cn.hutool.core.io.FileUtil;
+import com.datasophon.common.Constants;
+import com.datasophon.common.utils.ExecResult;
+import com.datasophon.worker.hook.HookAction;
+import com.datasophon.worker.hook.HookContext;
+import com.datasophon.common.utils.PkgInstallPathUtils;
+import com.datasophon.worker.utils.TaskConstants;
+import lombok.Data;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.File;
+import java.nio.charset.Charset;
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Data
+public class ReplaceStrategy implements HookAction {
+
+    public static final String REPLACE_TYPE = "replace";
+
+    private String source;
+
+    private String regex;
+
+    private String replacement;
+
+    @Override
+    public String getType() {
+        return REPLACE_TYPE;
+    }
+
+    @Override
+    public ExecResult invoke(HookContext context) {
+        BeanUtil.fillBeanWithMap(context.getParams(), this, CopyOptions.create().ignoreError());
+        Logger logger = LoggerFactory.getLogger(
+                TaskConstants.createLoggerName(context.getServiceName(), context.getServiceRoleName(), getClass()));
+        String basePath = PkgInstallPathUtils.getInstallHome(context);
+
+        logger.info("开始执行资源策略:{}...", getType());
+        File file = new File(basePath + Constants.SLASH + source);
+        if (file.exists()) {
+            List<String> lines = FileUtil.readLines(file, Charset.defaultCharset())
+                    .stream().map(line -> line.replaceAll(regex, replacement)).collect(Collectors.toList());
+            FileUtil.writeLines(lines, file, Charset.defaultCharset(), false);
+        }
+        return ExecResult.success();
+    }
+}
