--- datasophon-worker/src/main/java/com/datasophon/worker/strategy/resource/AppendLineStrategy.java
@@ -1,48 +1,0 @@
-package com.datasophon.worker.strategy.resource;
-
-import com.datasophon.common.Constants;
-
-import java.io.File;
-import java.nio.charset.Charset;
-import java.util.List;
-
-import com.datasophon.common.utils.ExecResult;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import cn.hutool.core.io.FileUtil;
-import cn.hutool.core.util.ObjectUtil;
-
-@EqualsAndHashCode(callSuper = true)
-@Data
-public class AppendLineStrategy extends ResourceStrategy {
-    
-    public static final String APPEND_LINE_TYPE = "append_line";
-    
-    private String source;
-    
-    private Integer line;
-    
-    private String text;
-
-    @Override
-    public String type() {
-        return APPEND_LINE_TYPE;
-    }
-
-    @Override
-    public ExecResult exec() {
-        logger.info("开始执行资源策略:{}...", type());
-        File file = new File(basePath + Constants.SLASH + source);
-        if (file.exists() && ObjectUtil.isNotNull(line)) {
-            List<String> lines = FileUtil.readLines(file, Charset.defaultCharset());
-            if (lines.size() >= line) {
-                String lineText = lines.get(line - 1);
-                if (!lineText.equals(text)) {
-                    lines.add(line - 1, text);
-                }
-                FileUtil.writeLines(lines, file, Charset.defaultCharset(), false);
-            }
-        }
-        return ExecResult.success();
-    }
-}
